rework poission simulations
I commented the code and increased the sample size. I corrected the text, which has the sizes of the bias all wrong. I also experimented briefly with a new version of a figure that Holger and I discussed, but that figure didn't seem to show the bias as clearly. The code for the new figure is still in the script.