integrate holger's changes
Notes on Major Changes

1. Improved subtitle.
2. I don't think RE logit is a ML estimator, so I removed that example.
3. I added comments at the top of each paragraph noting what I think that paragraph is trying to accomplish.
4. Since each y_i has it's own \theta_i, I'm not sure we should consider them iid, so I took that bit out.
5. I changed the probit model illustration of the inv. prop. to logit, since the cdf is more accessible.
6. I decided to refer to the additional/excessive bias as "unnecessary bias." Excessive seems to have a too-strong conotation like "more that you can reasonably tolerate."
7. In the drastic example, I removed references to the population, since there isn't a population in the sense that I understand it. Instead there's just a model/DGP.

Suggested language:

- Researchers use the invariance principle to compute the ML estimate.
- Researchers use the average of simulations to compute a point estimate.

Carsey and Harden (2011) suggests (singluar--the paper suggests).

Does it make sense to abbreviate the simulation average estimate to "SA estimate" like we shorten maximum likelihood estimate to "ML" estimate.

I'm referring to the average of the simulated quantities of interest as the "simulation average". Should we say "average of simulations" instead?

Are researchers interest in "quantities of interest" or "the quantity of interest". I'm going with the plural for now.

When listing both estimates, list mle first and avg second.